+++ opentelemetry-exporter-google-cloud/tests/test_cloud_monitoring_exporter_integration.py
@@ -13,10 +13,6 @@
 # limitations under the License.
 
 
-import socket
-import subprocess
-import unittest
-
 import grpc
 from google.cloud.monitoring_v3 import MetricServiceClient
 from google.cloud.monitoring_v3.gapic.transports import (
@@ -31,29 +27,7 @@ from opentelemetry.sdk import metrics
 from opentelemetry.sdk.metrics.export import MetricRecord, MetricsExportResult
 from opentelemetry.sdk.metrics.export.aggregate import SumAggregator
 
-
-# TODO: #46
-# Refactor duplicated code found in test_integration_cloud_trace_exporter.py
-class BaseExporterIntegrationTest(unittest.TestCase):
-    def setUp(self):
-        self.project_id = "TEST-PROJECT"
-
-        # Find a free port to spin up our server at.
-        sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        sock.bind(("localhost", 0))
-        self.address = "localhost:" + str(sock.getsockname()[1])
-        sock.close()
-
-        # Start the mock server.
-        args = ["mock_server", "-address", self.address]
-        self.mock_server_process = subprocess.Popen(
-            args, stderr=subprocess.PIPE
-        )
-        # Block until the mock server starts (it will output the address after starting).
-        self.mock_server_process.stderr.readline()
-
-    def tearDown(self):
-        self.mock_server_process.kill()
+from test_common import BaseExporterIntegrationTest
 
 
 class TestCloudMonitoringSpanExporter(BaseExporterIntegrationTest):
