+++ test-common/MANIFEST.in
@@ -1,0 +1,6 @@
+graft src
+graft tests
+global-exclude *.pyc
+global-exclude *.pyo
+global-exclude __pycache__/*
+include MANIFEST.in
