+++ test-common/setup.cfg
@@ -1,0 +1,13 @@
+[metadata]
+name = test-common
+description = Test utilities used across packages in the this repo
+
+[options]
+package_dir=
+    =src
+packages=find_namespace:
+install_requires =
+
+[options.packages.find]
+where = src
+
