+++ test-common/src/test_common/__init__.py
@@ -1,0 +1,7 @@
+"""Test utilities used across packages
+
+This code is only used in other modules' tests
+"""
+
+
+from .base_exporter_integration_test import BaseExporterIntegrationTest
