+++ tox.ini
@@ -21,6 +21,7 @@ base_deps =
   -c dev-constraints.txt
   opentelemetry-api
   opentelemetry-sdk
+  -e {toxinidir}/test-common
 
 dev_basepython = python3.8
 dev_deps =
